Taylor Glass contribution
I added to the cdc-exercise by creating a synthetic dataset to mimic the real data. I generated several plots to explore how successful the synthetic data was at modeling the real data. I explored the same questions that were presented for the real data and found similar answers using the synthetic data with the exception of the HA and NA subtypes question.